pyproject.toml
[tool.poetry]
name = "python_fastapi_data_retrival_app"
version = "0.0.1"
description = "This project using python fastAPI framwork to retrive data from a databricks workspace (you can customise to similar databases)"
authors = ["Lishuai Jing <flishuai.jing@gmail.com>"]
repository = "https://github.com/Lishuaijing3/python-fastapi-data-retrival-app"
documentation = "https://Lishuaijing3.github.io/python-fastapi-data-retrival-app/"
readme = "README.md"
packages = [
  {include = "python_fastapi_data_retrival_app"}
]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.13"
dash = "^2.14.0"
pyodbc = "^5.0.1"
uvicorn = "^0.23.2"
fastapi = "^0.104.0"
pandas = "^2.1.1"
dash-bootstrap-components = "^1.5.0"
python-dotenv = "^1.0.0"
streamlit = "^1.27.2"
gradio = "^3.50.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
deptry = "^0.6.4"
mypy = "^0.981"
pre-commit = "^2.20.0"
tox = "^3.25.1"
ipykernel = "^6.25.2"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2"
mkdocs-material = "^8.5.10"
mkdocstrings = {extras = ["python"], version = "^0.19.0"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py37']
preview = true

[tool.mypy]
files = ["python_fastapi_data_retrival_app"]
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"

[tool.ruff]
target-version = "py37"
line-length = 120
fix = true
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["python_fastapi_data_retrival_app"]


